+++ internal/cmd/domain.go
@@ -24,11 +24,11 @@ import (
 	"github.com/sapcc/limesctl/internal/core"
 )
 
-// DomainCmd contains the command-line structure for the domain command.
-type DomainCmd struct {
-	List domainListCmd `cmd:"" help:"Display data for all the domains. Requires a cloud-admin token."`
-	Show domainShowCmd `cmd:"" help:"Display data for a specific domain. Requires a domain-admin token."`
-	Set  domainSetCmd  `cmd:"" help:"Change quota values for a specific domain. Requires a cloud-admin token."`
+// domainCmd contains the command-line structure for the domain command.
+type domainCmd struct {
+	List domainListCmd `cmd:"" help:"Display resource usage data for all the domains. Requires a cloud-admin token."`
+	Show domainShowCmd `cmd:"" help:"Display resource usage data for a specific domain. Requires a domain-admin token."`
+	Set  domainSetCmd  `cmd:"" help:"Change resource quota values for a specific domain. Requires a cloud-admin token."`
 }
 
 //nolint:lll
@@ -36,23 +36,26 @@ type domainClusterFlag struct {
 	ClusterID string `short:"c" name:"cluster" help:"Cluster ID. When this option is used, the domain must be identified by ID (name won't work)."`
 }
 
+///////////////////////////////////////////////////////////////////////////////
+// Domain list.
+
 type domainListCmd struct {
 	domainClusterFlag
-	requestFilterFlags
-	outputFormatFlags
-}
-
-// Validate implements the kong.Validatable interface.
-func (d *domainListCmd) Validate() error {
-	return d.outputFormatFlags.validate()
+	resourceFilterFlags
+	resourceOutputFmtFlags
 }
 
 func (d *domainListCmd) Run(clients *ServiceClients) error {
+	outputOpts, err := d.resourceOutputFmtFlags.validate()
+	if err != nil {
+		return err
+	}
+
 	res := domains.List(clients.limes, domains.ListOpts{
-		Cluster:  d.ClusterID,
-		Area:     d.Area,
-		Service:  d.Service,
-		Resource: d.Resource,
+		Cluster:   d.ClusterID,
+		Areas:     d.Areas,
+		Services:  d.Services,
+		Resources: d.Resources,
 	})
 	if res.Err != nil {
 		return errors.Wrap(res.Err, "could not get domain reports")
@@ -67,22 +70,22 @@ func (d *domainListCmd) Run(clients *ServiceClients) error {
 		return errors.Wrap(err, "could not extract domain reports")
 	}
 
-	return writeReports(d.outputFormatFlags, core.LimesDomainsToReportRenderer(limesReps)...)
+	return writeReports(outputOpts, core.LimesDomainsToReportRenderer(limesReps)...)
 }
+
+///////////////////////////////////////////////////////////////////////////////
+// Domain show.
 
 type domainShowCmd struct {
 	domainClusterFlag
-	requestFilterFlags
-	outputFormatFlags
+	resourceFilterFlags
+	resourceOutputFmtFlags
 
 	NameOrID string `arg:"" optional:"" help:"Name or ID of the domain. Required if using '--cluster' flag."`
 }
 
 // Validate implements the kong.Validatable interface.
 func (d *domainShowCmd) Validate() error {
-	if err := d.outputFormatFlags.validate(); err != nil {
-		return err
-	}
 	if d.ClusterID != "" && d.NameOrID == "" {
 		return errors.New("Domain ID is required when using the '--cluster' flag")
 	}
@@ -90,6 +93,11 @@ func (d *domainShowCmd) Validate() error {
 }
 
 func (d *domainShowCmd) Run(clients *ServiceClients) error {
+	outputOpts, err := d.resourceOutputFmtFlags.validate()
+	if err != nil {
+		return err
+	}
+
 	domainID := d.NameOrID
 	if d.ClusterID == "" {
 		var err error
@@ -100,10 +108,10 @@ func (d *domainShowCmd) Run(clients *ServiceClients) error {
 	}
 
 	res := domains.Get(clients.limes, domainID, domains.GetOpts{
-		Cluster:  d.ClusterID,
-		Area:     d.Area,
-		Service:  d.Service,
-		Resource: d.Resource,
+		Cluster:   d.ClusterID,
+		Areas:     d.Areas,
+		Services:  d.Services,
+		Resources: d.Resources,
 	})
 	if res.Err != nil {
 		return errors.Wrap(res.Err, "could not get domain report")
@@ -118,12 +126,16 @@ func (d *domainShowCmd) Run(clients *ServiceClients) error {
 		return errors.Wrap(err, "could not extract domain report")
 	}
 
-	return writeReports(d.outputFormatFlags, core.DomainReport{DomainReport: limesRep})
+	return writeReports(outputOpts, core.DomainReport{DomainReport: limesRep})
 }
 
+///////////////////////////////////////////////////////////////////////////////
+// Domain set.
+
+//nolint:lll
 type domainSetCmd struct {
 	domainClusterFlag
-	Quotas []string `short:"q" help:"New quotas values. Example: service/resource=120GiB."`
+	Quotas []string `short:"q" sep:"," help:"New quotas values. For relative quota adjustment, use one of the following operators: [+=, -=, *=, /=]. Example: service/resource=10GiB."`
 
 	NameOrID string `arg:"" optional:"" help:"Name or ID of the domain. Required if using '--cluster' flag."`
 }
@@ -167,6 +179,7 @@ func (d *domainSetCmd) Run(clients *ServiceClients) error {
 }
 
 ///////////////////////////////////////////////////////////////////////////////
+// Helper functions.
 
 func getDomainResourceQuotas(limesClient *gophercloud.ServiceClient, clusterID, domainID string) (resourceQuotas, error) {
 	rep, err := domains.Get(limesClient, domainID, domains.GetOpts{