+++ internal/core/project.go
@@ -24,13 +24,19 @@ import (
 type ProjectReport struct {
 	*limes.ProjectReport
 
+	HasRatesOnly bool
+
 	DomainID   string
 	DomainName string
 }
 
 // LimesProjectsToReportRenderer wraps the given limes.ProjectReport in a
 // ProjectReport and returns a []LimesReportRenderer.
-func LimesProjectsToReportRenderer(in []limes.ProjectReport, domainID, domainName string) []LimesReportRenderer {
+func LimesProjectsToReportRenderer(
+	in []limes.ProjectReport,
+	domainID, domainName string,
+	hasRatesOnly bool) []LimesReportRenderer {
+
 	out := make([]LimesReportRenderer, 0, len(in))
 	for _, rep := range in {
 		rep := rep
@@ -38,6 +44,7 @@ func LimesProjectsToReportRenderer(in []limes.ProjectReport, domainID, domainNam
 			ProjectReport: &rep,
 			DomainID:      domainID,
 			DomainName:    domainName,
+			HasRatesOnly:  hasRatesOnly,
 		})
 	}
 	return out
@@ -45,26 +52,58 @@ func LimesProjectsToReportRenderer(in []limes.ProjectReport, domainID, domainNam
 
 var csvHeaderProjectDefault = []string{"domain id", "project id", "service", "resource", "quota", "usage", "unit"}
 
-var csvHeaderProjectLong = []string{"domain id", "domain name", "project id", "project name", "area", "service",
-	"category", "resource", "quota", "burst quota", "usage", "physical usage", "burst usage", "unit", "scraped at (UTC)"}
+var csvHeaderProjectLong = []string{
+	"domain id", "domain name", "project id", "project name", "area", "service",
+	"category", "resource", "quota", "burst quota", "usage", "physical usage", "burst usage", "unit", "scraped at (UTC)",
+}
 
 // GetHeaderRow implements the LimesReportRenderer interface.
-func (p ProjectReport) getHeaderRow(csvFmt CSVRecordFormat) []string {
-	switch csvFmt {
+func (p ProjectReport) getHeaderRow(opts *OutputOpts) []string {
+	if p.HasRatesOnly {
+		return p.getRatesHeaderRow(opts)
+	}
+
+	switch opts.CSVRecFmt {
 	case CSVRecordFormatLong:
 		return csvHeaderProjectLong
 	case CSVRecordFormatNames:
 		h := csvHeaderProjectDefault
-		h[0] = "domain name"
+		h[0] = domainName
 		h[1] = "project name"
 		return h
 	default:
+
 		return csvHeaderProjectDefault
 	}
 }
 
+var csvHeaderProjectRatesDefault = []string{"domain id", "project id", "service", "rate", "limit", "window", "usage", "unit"}
+
+var csvHeaderProjectRatesLong = []string{
+	"domain id", "domain name", "project id", "project name", "area", "service",
+	"rate", "limit", "default limit", "window", "default window", "usage", "unit", "scraped at (UTC)",
+}
+
+func (p ProjectReport) getRatesHeaderRow(opts *OutputOpts) []string {
+	switch opts.CSVRecFmt {
+	case CSVRecordFormatLong:
+		return csvHeaderProjectRatesLong
+	case CSVRecordFormatNames:
+		h := csvHeaderProjectRatesDefault
+		h[0] = domainName
+		h[1] = "project name"
+		return h
+	default:
+		return csvHeaderProjectRatesDefault
+	}
+}
+
 // Render implements the LimesReportRenderer interface.
-func (p ProjectReport) render(csvFmt CSVRecordFormat, humanize bool) CSVRecords {
+func (p ProjectReport) render(opts *OutputOpts) CSVRecords {
+	if p.HasRatesOnly {
+		return p.renderRates(opts)
+	}
+
 	var records CSVRecords
 
 	// Serialize service types with ordered keys
@@ -106,11 +145,11 @@ func (p ProjectReport) render(csvFmt CSVRecordFormat, humanize bool) CSVRecords 
 				}
 			}
 
-			valToStr, unit := getValToStrFunc(humanize, pSrvRes.Unit, []uint64{
+			valToStr, unit := getValToStrFunc(opts.Humanize, pSrvRes.Unit, []uint64{
 				zeroIfNil(burstQuota), burstUsage, zeroIfNil(physU), zeroIfNil(quota), usage,
 			})
 
-			if csvFmt == CSVRecordFormatLong {
+			if opts.CSVRecFmt == CSVRecordFormatLong {
 				r = append(r, p.DomainID, p.DomainName, p.UUID, p.Name, pSrv.Area, pSrv.Type, pSrvRes.Category,
 					pSrvRes.Name, emptyStrIfNil(quota, valToStr), emptyStrIfNil(burstQuota, valToStr), valToStr(usage),
 					emptyStrIfNil(physU, valToStr), valToStr(burstUsage), string(unit), timestampToString(pSrv.ScrapedAt),
@@ -118,12 +157,69 @@ func (p ProjectReport) render(csvFmt CSVRecordFormat, humanize bool) CSVRecords 
 			} else {
 				projectNameOrID := p.UUID
 				domainNameOrID := p.DomainID
-				if csvFmt == CSVRecordFormatNames {
+				if opts.CSVRecFmt == CSVRecordFormatNames {
 					projectNameOrID = p.Name
 					domainNameOrID = p.DomainName
 				}
 				r = append(r, domainNameOrID, projectNameOrID, pSrv.Type, pSrvRes.Name,
 					emptyStrIfNil(quota, valToStr), valToStr(usage), string(unit),
+				)
+			}
+
+			records = append(records, r)
+		}
+	}
+
+	return records
+}
+
+func (p ProjectReport) renderRates(opts *OutputOpts) CSVRecords {
+	var records CSVRecords
+
+	// Serialize service types with ordered keys
+	types := make([]string, 0, len(p.Services))
+	for typeStr := range p.Services {
+		types = append(types, typeStr)
+	}
+	sort.Strings(types)
+
+	for _, srv := range types {
+		// Serialize rate names with ordered keys
+		names := make([]string, 0, len(p.Services[srv].Rates))
+		for nameStr := range p.Services[srv].Rates {
+			names = append(names, nameStr)
+		}
+		sort.Strings(names)
+
+		for _, rate := range names {
+			var r []string
+			// Initialize temporary variables to make map lookups easier.
+			pSrv := p.Services[srv]
+			pSrvRate := p.Services[srv].Rates[rate]
+
+			var window, defaultWindow string
+			if pSrvRate.Window != nil {
+				window = pSrvRate.Window.String()
+			}
+			if pSrvRate.DefaultWindow != nil {
+				defaultWindow = pSrvRate.DefaultWindow.String()
+			}
+
+			valToStr := defaultValToStrFunc
+			if opts.CSVRecFmt == CSVRecordFormatLong {
+				r = append(r, p.DomainID, p.DomainName, p.UUID, p.Name, pSrv.Area, pSrv.Type, pSrvRate.Name,
+					valToStr(pSrvRate.Limit), valToStr(pSrvRate.DefaultLimit), window, defaultWindow,
+					pSrvRate.UsageAsBigint, string(pSrvRate.Unit), timestampToString(pSrv.RatesScrapedAt),
+				)
+			} else {
+				projectNameOrID := p.UUID
+				domainNameOrID := p.DomainID
+				if opts.CSVRecFmt == CSVRecordFormatNames {
+					projectNameOrID = p.Name
+					domainNameOrID = p.DomainName
+				}
+				r = append(r, domainNameOrID, projectNameOrID, pSrv.Type, pSrvRate.Name,
+					valToStr(pSrvRate.Limit), window, pSrvRate.UsageAsBigint, string(pSrvRate.Unit),
 				)
 			}